Add Medium/8/8.py solution

+++ Medium/7/7.py
@@ -16,9 +16,9 @@ class Solution:
         ans = 0
         while x != 0:
             digit = cpp_mod(x, 10)
-            if ans > INT_MAX / 10 or (ans == INT_MAX / 10 and digit > 7):
+            if ans > INT_MAX / 10 or (ans == int(INT_MAX / 10) and digit > 7):
                 return 0
-            if ans < INT_MIN / 10 or (ans == INT_MIN / 10 and digit < -8):
+            if ans < INT_MIN / 10 or (ans == int(INT_MIN / 10) and digit < -8):
                 return 0
             ans = ans * 10 + digit
             x = int(x / 10)
